+++ demos/customed/test_np.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+np_random.uniform(
+            low=[0, 0], high=[self.width, self.height], size=(self.env_ndim,)
+        )